+++ src/storage.ts
@@ -1,0 +1,9 @@
+import * as vscode from "vscode";
+
+export enum Location {
+  LSP = "lsp",
+}
+
+export function uri(location: Location, context: vscode.ExtensionContext): vscode.Uri {
+  return vscode.Uri.joinPath(context.globalStorageUri, location);
+}