+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 - Add syntax highlight for `meson.options` files.
 - Fix listing tests in the Test Explorer on activation
 - Add support to run executables that are not installed (eg: examples)
+- Add integration with
+  [Swift-MesonLSP](https://github.com/JCWasmx86/Swift-MesonLSP). vscode-meson
+  will automatically download it on Windows, Linux and MacOS. Linux users can
+  follow the [docs](https://github.com/JCWasmx86/Swift-MesonLSP/tree/main/Docs)
+  if they want to have updates delivered using their package manager.
 
 ## 1.10.0